[💎 : refactor] 파일 네이밍 컨벤션으로 케밥 케이스 적용
- 8bf37ba 작업에서 누락 부분 수행
