+++ src/components/tweet.tsx
@@ -3,7 +3,7 @@ import { deleteDoc, doc, getDoc } from 'firebase/firestore';
 import { deleteObject, ref } from 'firebase/storage';
 import { auth, db, storage } from '../firebase.ts';
 import ITweet from '../interfaces/ITweet.ts';
-import FormattedDate from '../utils/formattedDate.tsx';
+import FormattedDate from '../utils/formatted-date.tsx';
 import EditTweetForm from './edit-tweet-form.tsx';
 import * as S from '../styles/tweet.ts';
 import * as P from '../styles/popup.ts';
